Extract decodeTimestamp method
